rgx_j721*: sysconfig: enable active power management
Drop the power island control. The power island is controlled directly
by the firmware through the use of an MMR. Registering this domain with
the Linux power system interferes with this behavior.

Add active power managment hooks using the same logic as am62 now that
the device effectivly has 1 power domain under Linux.

This patch is dependent on device tree changes in Linux to drop the
second power domain for 8XE/BXS nodes on J* devices.

Signed-off-by: Randolph Sapp <rs@ti.com>