rgx_j721*: sysconfig: fix the power island definition
Apparently if bEnableRDPowIsland is enabled the core will expect some
form of feedback before the core is shutdown. Because this interaction
is currently occurring through a side band this feedback is never
received and a watchdog timeout is triggered. Add a note to make sure
nobody else misunderstands this feature.

Signed-off-by: Randolph Sapp <rs@ti.com>